pyproject.toml
[project]
name = "paperbanana"
version = "0.1.0"
description = "LangGraph implementation of the PaperBanana paper"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.41",
    "langchain-openai>=0.3.7",
    "langgraph>=0.2.67",
    "matplotlib>=3.10.0",
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "typer>=0.15.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[project.scripts]
paperbanana = "paperbanana.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["paperbanana*"]
